Up: split stuff into modules for testability, and install RSpec
