+++ Dockerfile
@@ -34,6 +34,6 @@ RUN gem install bundler
 COPY Gemfile Gemfile.lock /usr/src/
 RUN bundle install
 
-COPY gcloud.rb lib /usr/src/
+COPY app /usr/src/
 COPY config.yml /usr/src
-ENTRYPOINT ["bash", "-lc", "bundle exec ruby gcloud.rb"]
+ENTRYPOINT ["bash", "-lc", "bundle exec ruby turbulence.rb"]
