+++ Gemfile
@@ -1,3 +1,12 @@
+# frozen_string_literal: true
+
 source 'https://rubygems.org'
 
 gem 'tty-prompt', git: 'https://github.com/seawolf/tty-prompt.git', branch: 'auto-select-only-option'
+gem 'zeitwerk', '~> 2.4'
+
+group :development, :test do
+  gem 'rspec'
+  gem 'rubocop'
+  gem 'rubocop-rspec'
+end
