+++ turbulence.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require 'zeitwerk'
+
+Zeitwerk::Loader.new.tap do |loader|
+  loader.push_dir('app')
+
+  loader.inflector.inflect 'gcloud' => 'GCloud'
+
+  loader.enable_reloading
+  loader.setup
+  loader.eager_load
+end
